- missed these, the need to be ints due to the support library change

+++ frag-nav/src/main/java/com/ncapdevi/fragnav/tabhistory/FragNavTabHistoryController.kt
@@ -8,7 +8,7 @@ interface FragNavTabHistoryController {
     /**
      * Define what happens when we try to pop on a tab where root fragment is at the top
      */
-    @IntDef(CURRENT_TAB.toLong(), UNIQUE_TAB_HISTORY.toLong(), UNLIMITED_TAB_HISTORY.toLong())
+    @IntDef(CURRENT_TAB, UNIQUE_TAB_HISTORY, UNLIMITED_TAB_HISTORY)
     @kotlin.annotation.Retention(AnnotationRetention.SOURCE)
     annotation class PopStrategy
